# perf: improve tuple validation by precomputing TTUs