## Description
Closes https://github.com/openfga/openfga/issues/1104