@miparnisari 2023-11-24
@adriantam it's in the Pr summary

<img width="1264" alt="image" src="https://github.com/openfga/openfga/assets/5374887/5eb37437-aa7a-46c0-baf2-1fb70b9b6bf1">

The last column is the one that matters, less than 1 means improvement over `main`, more than 1 means regression.

🤔 I need a new benchmark that runs lots of tuple validations, which is where I expect this PR would show an improvement...