@miparnisari 2023-11-24
@adriantam  it looks like double as fast based on number of iterations made :D 

```
On main:

[24/11/23 11:43:47] ~/GitHub/openfga (main) $ go test ./internal/validation -bench=BenchmarkValidateTuple  -cpu 1 -benchmem -benchtime 5s -timeout 0 -run=^a
goos: darwin
goarch: amd64
pkg: github.com/openfga/openfga/internal/validation
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkValidateTuple   1292979              4639 ns/op            1232 B/op         11 allocs/op
PASS
ok      github.com/openfga/openfga/internal/validation  11.159s
[24/11/23 11:45:04] ~/GitHub/openfga (main) $ 


This PR:

[24/11/23 11:45:22] ~/GitHub/openfga (perf-improv) $  go test ./internal/validation -bench=BenchmarkValidateTuple  -cpu 1 -benchmem -benchtime 5s -timeout 0 -run=^a
goos: darwin
goarch: amd64
pkg: github.com/openfga/openfga/internal/validation
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkValidateTuple   2022118              3015 ns/op             120 B/op          3 allocs/op
PASS
ok      github.com/openfga/openfga/internal/validation  9.404s

```